requirements.txt
pdfplumber~=0.11.7
PyQt5~=5.15.11
python-docx~=1.2.0
python-pptx~=1.0.2